move simulator entrypoint to simulator/cmd/simulator
