+++ simulator/cmd/simulator/config/v1alpha1/register.go
